09-Authentication-and-Access-Bypass-Simulators/BypassBox/requirements.txt
pyjwt
colorama
requests
PyYAML
